1.1.0
lots of cleanup
refactoring
reorganizing/splitting
renamed to notify-util
now accessed via dot notation via default export
