+++ .prettierrc
@@ -1,5 +1,5 @@
 {
-  "printWidth": 80,
+  "printWidth": 90,
   "tabWidth": 2,
   "semi": true,
   "singleQuote": true,
