+++ src/config.ts
@@ -1,7 +1,9 @@
-import { autoModeCss, darkModeCss, defaultCss, injectCss } from './defaultCss';
 import { icons } from './icons';
+import { styles } from './styles';
+import { deepMerge } from './utils/deep-merge';
+import { injectCss } from './utils/inject-css';
 
-export const notificationConfig = {
+export const config = {
   containerSelector: '[nu_notification-container]',
   classes: {
     notificationClass: 'nu_notification',
@@ -19,58 +21,25 @@ export const notificationConfig = {
     spinner: icons.svg.spinner,
     info: icons.svg.info
   },
-  injectCss: true,
   theme: 'light'
 };
 
-// Import or define the DeepPartial type if you're using it inside the function for any reason.
-type DeepPartial<T> = {
-  [P in keyof T]?: T[P] extends object ? DeepPartial<T[P]> : T[P];
-};
-
-export function configureNotifications(options: any) {
-  // Runtime checks and merging logic here
-  const safeOptions = validateAndMergeOptions(options, notificationConfig);
-
+export function configure(options: any) {
   // Apply the safeOptions to your notificationConfig
-  deepMerge(notificationConfig, safeOptions);
+  deepMerge(config, options);
 
-  if (notificationConfig.injectCss) {
-    injectCss(defaultCss);
-    if (notificationConfig.theme === 'dark') {
-      injectCss(darkModeCss);
-      document.body.classList.add('nu_dark-mode');
-    }
-    if (notificationConfig.theme === 'auto') {
-      injectCss(autoModeCss);
-    }
+  // Adds the appropriate styles based on the theme
+  switch (config.theme) {
+    case 'none':
+      break;
+    case 'light':
+      injectCss(styles.light, styles.base);
+      break;
+    case 'dark':
+      injectCss(styles.dark, styles.base);
+      break;
+    case 'auto':
+      injectCss(styles.auto, styles.base);
+      break;
   }
-}
-
-function validateAndMergeOptions(
-  userOptions: any,
-  defaults: typeof notificationConfig
-): DeepPartial<typeof notificationConfig> {
-  // Implement your validation and merging logic here
-  // This is a placeholder function to illustrate the concept
-  // You would need to recursively validate and merge the userOptions with your defaults
-  return userOptions; // This should be the result of your actual merging logic
-}
-
-// Existing deepMerge function or similar logic to apply the safeOptions
-function deepMerge(target: any, source: any) {
-  Object.keys(source).forEach((key) => {
-    if (
-      source[key] &&
-      typeof source[key] === 'object' &&
-      !(source[key] instanceof Array)
-    ) {
-      if (!target[key]) {
-        target[key] = {};
-      }
-      deepMerge(target[key], source[key]);
-    } else {
-      target[key] = source[key];
-    }
-  });
 }
