+++ src/icons/get-icon.ts
@@ -1,33 +1,33 @@
-import { icons } from '.';
-import { notificationConfig } from '../config';
-
 /**
  * Retrieves the HTML string for a given icon type.
+ * This method allows the user to set their own icons, overwriting the default config icons.
  *
  * @param {string} type - The type of icon to retrieve. Valid types are 'success', 'debug', 'error', 'warning', and 'loading'.
  * @returns {string} The HTML string representing the requested icon.
  */
+import { config } from '../config';
+
 export function getIcon(type: string): string {
   let iconHTML: string;
 
   switch (type) {
     case 'success':
-      iconHTML = notificationConfig.icons.success;
+      iconHTML = config.icons.success;
       break;
     case 'debug':
-      iconHTML = notificationConfig.icons.debug;
+      iconHTML = config.icons.debug;
       break;
     case 'error':
-      iconHTML = notificationConfig.icons.error;
+      iconHTML = config.icons.error;
       break;
     case 'warning':
-      iconHTML = notificationConfig.icons.warning;
+      iconHTML = config.icons.warning;
       break;
     case 'spinner':
-      iconHTML = notificationConfig.icons.spinner;
+      iconHTML = config.icons.spinner;
       break;
     case 'info':
-      iconHTML = notificationConfig.icons.info;
+      iconHTML = config.icons.info;
       break;
   }
