+++ src/icons/spinner.ts
@@ -1,1 +1,2 @@
+// This one is accomplished with CSS, see `../styles/base.ts` -> `.nu_spinner`
 export const spinnerIcon = `<div></div>`;
