--- src/loader/index.ts
@@ -1,7 +1,0 @@
-import { startLoader } from './start-loader';
-import { updateLoader } from './update-loader';
-
-export const loader = {
-  start: startLoader,
-  update: updateLoader
-};
