+++ src/notifications/default-options.ts
@@ -1,0 +1,7 @@
+// Default options
+export const defaultOptions = {
+  heading: '', // Required
+  body: '',
+  duration: 3500, // in ms, null for infinite
+  clickToClose: true
+};
