+++ src/styles/dark.ts
@@ -1,0 +1,26 @@
+export const darkStyles = `
+:root {
+  --nu--background: var(--nu-slate-800);
+  --nu--text: var(--nu-slate-300);
+  --nu--heading: var(--nu-slate-50);
+  --nu--border: var(--nu-slate-700);
+  --nu-progress-bar: var(--nu-slate-500);
+
+  --nu-success-icon: var(--nu-green-500);
+  --nu-success-border: var(--nu-green-600);
+
+  --nu-error-icon: var(--nu-red-500);
+  --nu-error-border: var(--nu-red-500);
+
+  --nu-warning-icon: var(--nu-yellow-500);
+  --nu-warning-border: var(--nu-yellow-400);
+
+  --nu-info-icon: var(--nu-slate-400);
+  --nu-info-border: var(--nu-slate-700);
+
+  --nu-debug-icon: var(--nu-blue-300);
+  --nu-debug-border: var(--nu-blue-600);
+
+  --nu-spinner-icon: var(--nu-slate-400);
+  --nu-spinner-border: var(--nu-slate-400);
+}`;
