+++ src/styles/index.ts
@@ -1,0 +1,11 @@
+import { autoStyles } from './auto';
+import { baseStyles } from './base';
+import { darkStyles } from './dark';
+import { lightStyles } from './light';
+
+export const styles = {
+  base: baseStyles,
+  light: lightStyles,
+  dark: darkStyles,
+  auto: autoStyles
+};
