+++ src/styles/light.ts
@@ -1,0 +1,27 @@
+export const lightStyles = `
+:root {
+  --nu--background: var(--nu-white);
+  --nu--text: var(--nu-slate-600);
+  --nu--heading: var(--nu-slate-800);
+  --nu--border: var(--nu-slate-200);
+  --nu-progress-bar: var(--nu-slate-300);
+
+  --nu-success-icon: var(--nu-green-600);
+  --nu-success-border: var(--nu-green-600);
+
+  --nu-error-icon: var(--nu-red-500);
+  --nu-error-border: var(--nu-red-500);
+
+  --nu-warning-icon: var(--nu-yellow-500);
+  --nu-warning-border: var(--nu-yellow-400);
+
+  --nu-info-icon: var(--nu-slate-500);
+  --nu-info-border: var(--nu-slate-200);
+
+  --nu-debug-icon: var(--nu-blue-600);
+  --nu-debug-border: var(--nu-blue-600);
+
+  --nu-spinner-icon: var(--nu-slate-500);
+  --nu-spinner-border: var(--nu-slate-400);
+}
+`;
