+++ src/utils/validate-and-merge-options.ts
@@ -1,0 +1,27 @@
+import { config } from '../config';
+
+/**
+ * Represents a recursive partial type.
+ * Each property of the given type is made optional and any properties that are objects are also made recursive partials.
+ * @template T - The type to be made into a DeepPartial.
+ */
+type DeepPartial<T> = {
+  [P in keyof T]?: T[P] extends object ? DeepPartial<T[P]> : T[P];
+};
+
+/**
+ * Validates and merges user-provided options with default configuration options.
+ *
+ * @param {any} userOptions - The options provided by the user.
+ * @param {typeof config} defaults - The default configuration options.
+ * @returns {DeepPartial<typeof config>} The merged options after validation.
+ */
+export function validateAndMergeOptions(
+  userOptions: any,
+  defaults: typeof config
+): DeepPartial<typeof config> {
+  // Implement your validation and merging logic here
+  // This is a placeholder function to illustrate the concept
+  // You would need to recursively validate and merge the userOptions with your defaults
+  return userOptions; // This should be the result of your actual merging logic
+}
