+++ templates/style.css
@@ -27,7 +27,7 @@ body {
   --nu--text: var(--nu-slate-600);
   --nu--heading: var(--nu-slate-800);
   --nu--border: var(--nu-slate-200);
-  --nu-progress-bar: var(--nu-slate-200);
+  --nu-progress-bar: var(--nu-slate-300);
 
   --nu-success-icon: var(--nu-green-600);
   --nu-success-border: var(--nu-green-600);
@@ -39,7 +39,7 @@ body {
   --nu-warning-border: var(--nu-yellow-400);
 
   --nu-info-icon: var(--nu-slate-500);
-  --nu-info-border: var(--nu-slate-300);
+  --nu-info-border: var(--nu-slate-200);
 
   --nu-debug-icon: var(--nu-blue-600);
   --nu-debug-border: var(--nu-blue-600);
